refactor: wgpu and memory efficiency experiments (#7)
* chore: delete bmps

* chore: clean up

* fix: recreate vertex buffer when increasing sprites

* refactor: wgpu experiments and memory efficient

+++ crates/definitions/src/types/animation.rs
@@ -2,5 +2,5 @@
 pub struct Animation {
     pub id: usize,
     pub speed: usize,
-    pub frames: Vec<usize>,
+    pub frames: Vec<u32>,
 }
